Move the rest of the Cohost folder up to Razor projects, and flatten startup service factory
