+++ src/Razor/src/Razor/src/Microsoft.CodeAnalysis.Razor.CohostingShared/CohostEndpointAttribute.cs
@@ -1,7 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Microsoft.CodeAnalysis.ExternalAccess.Razor.Cohost;
+using Microsoft.CodeAnalysis.LanguageServer;
 using Microsoft.CommonLanguageServerProtocol.Framework;
 
 namespace Microsoft.VisualStudio.Razor.LanguageClient.Cohost;
@@ -9,7 +9,7 @@ namespace Microsoft.VisualStudio.Razor.LanguageClient.Cohost;
 internal class CohostEndpointAttribute : LanguageServerEndpointAttribute
 {
     public CohostEndpointAttribute(string method)
-        : base(method, Constants.RazorLanguageName)
+        : base(method, LanguageInfoProvider.RazorLanguageName)
     {
     }
 }
