+++ src/Razor/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Diagnostics/RazorTranslateDiagnosticsService.cs
@@ -17,6 +17,7 @@ using Microsoft.CodeAnalysis.Razor.DocumentMapping;
 using Microsoft.CodeAnalysis.Razor.Logging;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 using Microsoft.CodeAnalysis.Razor.Protocol;
+using Microsoft.CodeAnalysis.RemoveUnnecessaryImports;
 using Microsoft.CodeAnalysis.Text;
 
 namespace Microsoft.CodeAnalysis.Razor.Diagnostics;
@@ -487,9 +488,9 @@ internal class RazorTranslateDiagnosticsService(IDocumentMappingService document
             if (str switch
             {
                 "CS1525" => ShouldIgnoreCS1525(diagnostic, codeDocument),
-                Constants.DiagnosticIds.IDE0005_gen => IsUsingDirectiveUsed(diagnostic, codeDocument),
+                RemoveUnnecessaryImportsConstants.IDE0005_gen => IsUsingDirectiveUsed(diagnostic, codeDocument),
                 // This diagnostics is produced by Roslyn to help its Remove Usings code fixer, so is irrelevant to us
-                Constants.DiagnosticIds.RemoveUnnecessaryImportsFixable => true,
+                RemoveUnnecessaryImportsConstants.DiagnosticFixableId => true,
                 _ => false
             })
             {
