+++ src/Razor/src/Razor/src/Microsoft.CodeAnalysis.Remote.Razor/Diagnostics/RemoteDiagnosticsService.cs
@@ -13,8 +13,8 @@ using Microsoft.CodeAnalysis.Razor.Protocol;
 using Microsoft.CodeAnalysis.Razor.Remote;
 using Microsoft.CodeAnalysis.Razor.Workspaces.Settings;
 using Microsoft.CodeAnalysis.Remote.Razor.ProjectSystem;
+using Microsoft.CodeAnalysis.RemoveUnnecessaryImports;
 using Microsoft.CodeAnalysis.Text;
-using EAConstants = Microsoft.CodeAnalysis.ExternalAccess.Razor.Cohost.Constants;
 
 namespace Microsoft.CodeAnalysis.Remote.Razor;
 
@@ -77,7 +77,7 @@ internal sealed class RemoteDiagnosticsService(in ServiceArgs args) : RazorDocum
 
         foreach (var diagnostic in allDiagnostics)
         {
-            if (diagnostic.Code is { Value: EAConstants.DiagnosticIds.IDE0005_gen })
+            if (diagnostic.Code is { Value: RemoveUnnecessaryImportsConstants.IDE0005_gen })
             {
                 var absoluteIndex = sourceText.GetRequiredAbsoluteIndex(diagnostic.Range.Start.Line, diagnostic.Range.Start.Character);
                 var token = tree.Root.FindToken(absoluteIndex);
@@ -132,7 +132,7 @@ internal sealed class RemoteDiagnosticsService(in ServiceArgs args) : RazorDocum
                 diagnostics.Add(new LspDiagnostic
                 {
                     // We log the same as Roslyn does, so we can have only one post-report cleanup pass, above.
-                    Code = EAConstants.DiagnosticIds.IDE0005_gen,
+                    Code = RemoveUnnecessaryImportsConstants.IDE0005_gen,
                     Message = SR.AddTagHelper_directive_is_unnecessary,
                     Source = LanguageServerConstants.RazorDiagnosticSource,
                     Range = sourceText.GetRange(directive.SpanWithoutTrailingNewLines(sourceText)),
